services/pet-detection/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
requests
python-multipart
numpy
scikit-learn
